requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.4.0
django-allauth[socialaccount]==66.0.2

# Database (PostgreSQL)
psycopg[binary]>=3.2.0

# Celery + Redis
celery==5.4.0
django-celery-results==2.5.1
redis==5.2.1

# API Integrations
requests==2.32.3
openai
google-generativeai

# Utilities
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
